Update solution.
Change working WIL version to a newer version.
In the project file for MemoryCommando.Tests project, make path of a reference a relative path to the project instead of a solution: to make it work as submodules in other solutions.
Disable 4996 warning.
Make WIL NuGet package to be relative to the solution directory instead of a project directory.